requirements.txt
python-telegram-bot==21.10
Django==4.2.18
psycopg2-binary>=2.9,<3.0
gunicorn==23.0.0
memcache==0.12.0
pymemcache==4.0.0
